refactor: improve bath transaction support
Also did major refactoring on core class imports. Every base class, entity or procedure must be now imported into a `~/internal` file and exported from there.
This is to ensure a specific load order and prevent nasty circular dependency issues.
More info on the pattern: https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de