fix(documents): format documents correctly when adding them to a token
The underlying format in the chain had changed but we missed it

BREAKING CHANGE: Make `contentHash` in the `TokenDocument` interface optional. This affects the
output of `token.documents.get`. An error is thrown when adding documents to a token if the content
hash is not a hex string with a 0x prefix