fix: 🐛 Remove polymeshApi Proxy in Context
polkadot.js began using a private accesor (e.g. #registry) internally.
This causes an error to be thrown when called by a proxied instance.
Context Proxy was used to guard against having no signer which is taken
care of by the singning manager, so this shouldn't cause any behavior to
change