feat: 🎸 add support for padded IDs in subquery
Add dual version SubQuery support. v19+ SQ instance will pad block ID
values so ordering by created block is such that the lexical order is
equivalent to numerical order. Also created_at column has been removed
as the padding makes it redundant (it was the time of sync, not the time
on chain).